Fix sensor id read returning 0 if reset-gpio is missing
With new device trees, reset-gpio is usage is dropped for regulators
instead.

Prior to this, the sensor FSM was checked only when reset-gpio was
handled, thus missing the state wait for the sensor being ready,
effectively returning 0 on the first i2c read as it was too early for
the sensor.

Fix this by moving the wait state code outside of the gpio reset code,
performing the check unconditionally.

Signed-off-by: Benjamin Mugnier <benjamin.mugnier@foss.st.com>